fixes several import errors and other, there is still one issue with the deployment of commands
