+++ utils/utils.js
@@ -1,7 +1,6 @@
 import fetch from 'node-fetch';
 import { verifyKey } from 'discord-interactions';
 import { config } from '../config.js';
-const { LEETCODE_URL } = config;
 
 export function VerifyDiscordRequest(clientKey) {
   return function (req, res, buf, encoding) {
@@ -64,104 +63,4 @@ export function getRandomEmoji() {
 
 export function capitalize(str) {
   return str.charAt(0).toUpperCase() + str.slice(1);
-}
-
-// Leetcode helpers
-
-export async function dailyGetLeetcodeData() {
-  const data = {
-    query: `
-  query questionOfToday {
-      activeDailyCodingChallengeQuestion {
-        date
-        userStatus
-        link
-        question {
-          acRate
-          difficulty
-          freqBar
-          frontendQuestionId: questionFrontendId
-          isFavor
-          paidOnly: isPaidOnly
-          status
-          title
-          titleSlug
-          hasVideoSolution
-          hasSolution
-          topicTags {
-            name
-            id
-            slug
-          }
-        }
-      }
-    }`,
-    variables: {},
-  };
-
-  const response = await fetch(`${LEETCODE_URL}/graphql/`, {
-    method: 'POST',
-    body: JSON.stringify(data),
-    headers: {
-      'Content-Type': 'application/json',
-      'Accept-Encoding': 'gzip, deflate, br',
-    },
-    compress: true,
-  });
-
-  if (response.status !== 200) {
-    console.error('could not fetch: ', response.status);
-    return;
-  }
-  return response.json();
-}
-
-export async function weeklyGetLeetcodeData() {
-  const now = new Date();
-
-  const data = {
-    query: `
-  query dailyCodingQuestionRecords($year: Int!, $month: Int!) {
-      dailyCodingChallengeV2(year: $year, month: $month) {
-          challenges {
-          date
-          userStatus
-          link
-          question {
-              questionFrontendId
-              title
-              titleSlug
-          }
-          }
-          weeklyChallenges {
-          date
-          userStatus
-          link
-          question {
-              questionFrontendId
-              title
-              titleSlug
-          }
-          }
-      }
-  }`,
-    variables: { year: now.getFullYear(), month: now.getMonth() + 1 },
-  };
-
-  const response = await fetch(`${LEETCODE_URL}/graphql/`, {
-    method: 'POST',
-    body: JSON.stringify(data),
-    headers: {
-      'Content-Type': 'application/json',
-      'Accept-Encoding': 'gzip, deflate, br',
-    },
-    compress: true,
-  });
-
-  if (response.status !== 200) {
-    console.error('could not fetch: ', response.status);
-    return;
-  }
-
-  return response.json();
 }
